flask-app/requirements.txt
flask 
psycopg2-binary 
flask-login